Applied Ivan comments
Updated tests

Added Broadcast v3 support

Moved pass into separate file
