+++ src/common/snippets/src/pass/align_element_type.cpp
@@ -29,7 +29,8 @@ inline auto op_supports_only_exec_type(const std::shared_ptr<ov::Node>& n) -> bo
            !ov::is_type<ov::op::v0::Result>(n) &&
            !ov::is_type<ov::op::v1::Transpose>(n) &&
            !ov::is_type<ov::op::v0::Convert>(n) &&
-           !ov::is_type<ov::op::v1::Broadcast>(n);
+           !ov::is_type<ov::op::v1::Broadcast>(n) &&
+           !ov::is_type<ov::op::v3::Broadcast>(n);
 }
 
 }  // namespace
