+++ src/common/snippets/tests/include/pass/insert_movebroadcast.hpp
@@ -22,24 +22,10 @@ typedef std::tuple<
         Shape // Broadcast shape 1
 > insertMoveBroadcastParams;
 
-typedef std::tuple<
-        Shape, // Input shape 0
-        Shape, // Input shape 1
-        Shape  // Broadcast shape
-> BroadcastParams;
-
 using ngraph::snippets::op::Subgraph;
 class InsertMoveBroadcastTests : public LoweringTests, public testing::WithParamInterface<insertMoveBroadcastParams> {
 public:
     static std::string getTestCaseName(testing::TestParamInfo<insertMoveBroadcastParams> obj);
-protected:
-    void SetUp() override;
-    std::shared_ptr<SnippetsFunctionBase> snippets_function;
-};
-
-class BroadcastToMoveBroadcastTests : public LoweringTests, public testing::WithParamInterface<BroadcastParams> {
-public:
-    static std::string getTestCaseName(testing::TestParamInfo<BroadcastParams> obj);
 protected:
     void SetUp() override;
     std::shared_ptr<SnippetsFunctionBase> snippets_function;
