+++ src/plugins/intel_cpu/tests/functional/shared_tests_instances/snippets/select.cpp
@@ -18,7 +18,7 @@ INSTANTIATE_TEST_SUITE_P(smoke_Snippets_Select, Select,
                 ::testing::ValuesIn({ov::Shape{1, 5, 5, 35}, ov::Shape{1}}),
                 ::testing::ValuesIn({ov::Shape{1, 5, 5, 35}, ov::Shape{1}}),
                 ::testing::ValuesIn({ov::element::f32, ov::element::i8}),
-                ::testing::Values(4), // 1 Subgraph + 2 Sinh + 1 Roll
+                ::testing::Values(1),
                 ::testing::Values(1),
                 ::testing::Values(CommonTestUtils::DEVICE_CPU)),
         Select::getTestCaseName);
@@ -30,7 +30,7 @@ INSTANTIATE_TEST_SUITE_P(smoke_Snippets_BroadcastSelect, BroadcastSelect,
                                  ::testing::ValuesIn({Shape{1, 8, 2, 10}, Shape{1, 1, 1, 1}}),
                                  ::testing::ValuesIn({Shape{1, 8, 2, 1}, Shape{1, 8, 2, 10}}),
                                  ::testing::ValuesIn({ov::element::f32, ov::element::i8}),
-                                 ::testing::Values(4), // 1 Subgraph + 2 Sinh + 1 Roll
+                                 ::testing::Values(1),
                                  ::testing::Values(1),
                                  ::testing::Values(CommonTestUtils::DEVICE_CPU)),
                          BroadcastSelect::getTestCaseName);
