+++ src/tests/ngraph_helpers/snippets_ngraph_functions/include/subgraph_softmax.hpp
@@ -21,29 +21,9 @@ protected:
     int axis;
 };
 
-class SinhSoftmaxFunction : public SnippetsFunctionBase {
-public:
-    explicit SinhSoftmaxFunction(const std::vector<PartialShape>& inputShapes, int axis) : SnippetsFunctionBase(inputShapes), axis(axis) {
-        NGRAPH_CHECK(input_shapes.size() == 1, "Got invalid number of input shapes");
-    }
-protected:
-    std::shared_ptr<ov::Model> initOriginal() const override;
-    int axis;
-};
-
 class AddSoftmaxFunction : public SnippetsFunctionBase {
 public:
     explicit AddSoftmaxFunction(const std::vector<PartialShape>& inputShapes, int axis) : SnippetsFunctionBase(inputShapes), axis(axis) {
-        NGRAPH_CHECK(input_shapes.size() == 2, "Got invalid number of input shapes");
-    }
-protected:
-    std::shared_ptr<ov::Model> initOriginal() const override;
-    int axis;
-};
-
-class SinhAddSoftmaxFunction : public SnippetsFunctionBase {
-public:
-    explicit SinhAddSoftmaxFunction(const std::vector<PartialShape>& inputShapes, int axis) : SnippetsFunctionBase(inputShapes), axis(axis) {
         NGRAPH_CHECK(input_shapes.size() == 2, "Got invalid number of input shapes");
     }
 protected:
